backend/requirements/dev.txt
-r base.txt

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3
ipython==8.15.0
black==23.7.0
flake8==6.0.0
pytest==7.4.0
pytest-django==4.5.2
factory-boy==3.3.0
watchdog==3.0.0 